requirements.txt
# Core Dependencies
uagents>=0.22.0
uagents-core

# AI & ML
anthropic>=0.40.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async & Networking
asyncio>=3.4.3
aiohttp>=3.9.0
httpx>=0.27.0

# Environment & Config
python-dotenv>=1.0.0

# Date & Time
python-dateutil>=2.8.0

# Logging & Monitoring
structlog>=24.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Type Checking
mypy>=1.8.0
types-python-dateutil>=2.8.0

# Code Quality
black>=24.0.0
ruff>=0.2.0

# Utilities
uuid>=1.30
psutil>=5.9.0